PBI 39423 Save validation result.

+++ Infrastructure/ApplicationLayer/ScheduleProjectionReadOnly/ReadModelValidationResultPersister.cs
@@ -38,6 +38,18 @@ namespace Teleopti.Ccc.Infrastructure.ApplicationLayer.ScheduleProjectionReadOnl
 				.ExecuteUpdate();
 		}
 
+		public void SaveScheduleDay(ReadModelValidationResult input)
+		{
+			_currentUnitOfWork.Session()
+				.CreateSQLQuery(
+					@"INSERT INTO [ReadModel].[ScheduleDay_check] (PersonId, BelongsToDate, IsValid, UpdatedOn) VALUES (:PersonId, :BelongsToDate, :IsValid, :UpdatedOn)")
+				.SetDateTime("BelongsToDate",input.Date)
+				.SetGuid("PersonId",input.PersonId)
+				.SetBoolean("IsValid",input.IsValid)
+				.SetDateTime("UpdatedOn",DateTime.UtcNow)
+				.ExecuteUpdate();
+		}
+
 		public IEnumerable<ReadModelValidationResult> LoadAllInvalid()
 		{
 			var result = _currentUnitOfWork.Session().CreateSQLQuery(
@@ -47,10 +59,25 @@ namespace Teleopti.Ccc.Infrastructure.ApplicationLayer.ScheduleProjectionReadOnl
 			return result;
 		}
 
-		public void Reset()
+		public void Reset(IList<ValidateReadModelType> types)
 		{
-			_currentUnitOfWork.Session().CreateSQLQuery("TRUNCATE TABLE [ReadModel].[ScheduleProjectionReadOnly_check]")
-				.ExecuteUpdate();
+			if (types.Contains(ValidateReadModelType.ScheduleProjectionReadOnly))
+			{
+				_currentUnitOfWork.Session().CreateSQLQuery("TRUNCATE TABLE [ReadModel].[ScheduleProjectionReadOnly_check]")
+					.ExecuteUpdate();
+			}
+
+			if(types.Contains(ValidateReadModelType.PersonScheduleDay))
+			{
+				_currentUnitOfWork.Session().CreateSQLQuery("TRUNCATE TABLE [ReadModel].[PersonScheduleDay_check]")
+					.ExecuteUpdate();
+			}
+
+			if(types.Contains(ValidateReadModelType.ScheduleDay))
+			{
+				_currentUnitOfWork.Session().CreateSQLQuery("TRUNCATE TABLE [ReadModel].[ScheduleDay_check]")
+					.ExecuteUpdate();
+			}
 		}
 	}
 }
